Separate getShapecastFn
